+++ src/us/kbase/common/executionengine/JobRunnerConstants.java
@@ -1,4 +1,4 @@
-package us.kbase.narrativejobservice;
+package us.kbase.common.executionengine;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashSet;
