+++ src/us/kbase/narrativejobservice/test/CallbackServerTest.java
@@ -47,20 +47,20 @@ import us.kbase.auth.AuthService;
 import us.kbase.auth.AuthToken;
 import us.kbase.catalog.CatalogClient;
 import us.kbase.catalog.SelectOneModuleParams;
+import us.kbase.common.executionengine.CallbackServer;
+import us.kbase.common.executionengine.CallbackServerConfigBuilder;
+import us.kbase.common.executionengine.LineLogger;
+import us.kbase.common.executionengine.ModuleMethod;
+import us.kbase.common.executionengine.ModuleRunVersion;
+import us.kbase.common.executionengine.CallbackServerConfigBuilder.CallbackServerConfig;
 import us.kbase.common.service.JacksonTupleModule;
 import us.kbase.common.service.JsonClientException;
 import us.kbase.common.service.JsonServerServlet;
 import us.kbase.common.service.ServerException;
 import us.kbase.common.service.UObject;
 import us.kbase.common.test.controllers.ControllerCommon;
-import us.kbase.common.utils.ModuleMethod;
-import us.kbase.narrativejobservice.LineLogger;
 import us.kbase.narrativejobservice.NarrativeJobServiceServer;
-import us.kbase.narrativejobservice.subjobs.CallbackServer;
 import us.kbase.narrativejobservice.subjobs.NJSCallbackServer;
-import us.kbase.narrativejobservice.subjobs.CallbackServerConfigBuilder;
-import us.kbase.narrativejobservice.subjobs.ModuleRunVersion;
-import us.kbase.narrativejobservice.subjobs.CallbackServerConfigBuilder.CallbackServerConfig;
 import us.kbase.workspace.ProvenanceAction;
 import us.kbase.workspace.SubAction;
